refactor: CardPreviewInfoType 및 CardFormInfoType을 feature 폴더로 이동
